Add error handling and check for authorization/authentication in activities controller

+++ app/controllers/activities_controller.rb
@@ -2,12 +2,11 @@ class ActivitiesController < ApplicationController
     def create
         decoded_token = JsonWebToken.decode(request.headers['Authorization'])
         # user = User.find(decoded_token[:user_id])
-        # User not used now but might be useful for authorization?
         activity = Activity.new(activity_params)
         if activity.save
             render json: activity
         else
-            # send error
+            render json: { message: activity.errors }, status: 400
         end
     end
 
@@ -16,10 +15,15 @@ class ActivitiesController < ApplicationController
         decoded_token = JsonWebToken.decode(request.headers['Authorization'])
         user = User.find(decoded_token[:user_id])
         activity = Activity.find(params[:id])
-        activity.destroy
-        render json: { success: true }
-    end
 
+        if activity.trip.user === user
+            activity.destroy
+            render status: 204
+        else
+            render json: { message: "Unable to delete activity"}, status: 400
+        end
+
+    end
 
     private
